Break up external integration and internal unit tests
