+++ api/trace/b3_propagator_test.go
@@ -15,8 +15,6 @@
 package trace
 
 import (
-	"context"
-	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -28,265 +26,6 @@ var (
 	spanID     = SpanID{0, 0, 0, 0, 0, 0, 0, 0x7b}
 	spanIDStr  = "000000000000007b"
 )
-
-type Supplier struct {
-	SingleHeader       string
-	DebugFlagHeader    string
-	TraceIDHeader      string
-	SpanIDHeader       string
-	SampledHeader      string
-	ParentSpanIDHeader string
-}
-
-func (s *Supplier) Get(key string) string {
-	switch key {
-	case B3SingleHeader:
-		return s.SingleHeader
-	case B3DebugFlagHeader:
-		return s.DebugFlagHeader
-	case B3TraceIDHeader:
-		return s.TraceIDHeader
-	case B3SpanIDHeader:
-		return s.SpanIDHeader
-	case B3SampledHeader:
-		return s.SampledHeader
-	case B3ParentSpanIDHeader:
-		return s.ParentSpanIDHeader
-	}
-	return ""
-}
-
-func (s *Supplier) Set(key, value string) {
-	fmt.Println("called", key, value)
-	switch key {
-	case B3SingleHeader:
-		s.SingleHeader = value
-	case B3DebugFlagHeader:
-		s.DebugFlagHeader = value
-	case B3TraceIDHeader:
-		s.TraceIDHeader = value
-	case B3SpanIDHeader:
-		s.SpanIDHeader = value
-	case B3SampledHeader:
-		s.SampledHeader = value
-	case B3ParentSpanIDHeader:
-		s.ParentSpanIDHeader = value
-	}
-}
-
-type TestSpan struct {
-	NoopSpan
-	sc SpanContext
-}
-
-func (s TestSpan) SpanContext() SpanContext {
-	return s.sc
-}
-
-func TestInject(t *testing.T) {
-	tests := []struct {
-		sc       SpanContext
-		b3       B3
-		expected *Supplier
-	}{
-		{
-			sc:       SpanContext{},
-			b3:       B3{},
-			expected: &Supplier{},
-		},
-		{
-			sc: SpanContext{
-				SpanID:     spanID,
-				TraceFlags: FlagsSampled,
-			},
-			b3:       B3{},
-			expected: &Supplier{},
-		},
-		{
-			sc: SpanContext{
-				TraceID:    traceID,
-				TraceFlags: FlagsSampled,
-			},
-			b3:       B3{},
-			expected: &Supplier{},
-		},
-		{
-			sc: SpanContext{
-				TraceFlags: FlagsSampled,
-			},
-			b3:       B3{},
-			expected: &Supplier{},
-		},
-		{
-			sc: SpanContext{
-				TraceID:    traceID,
-				SpanID:     spanID,
-				TraceFlags: FlagsSampled,
-			},
-			b3: B3{SingleHeader: true},
-			expected: &Supplier{
-				SingleHeader: "000000000000007b00000000000001c8-000000000000007b-1",
-			},
-		},
-		{
-			sc: SpanContext{
-				TraceID: traceID,
-				SpanID:  spanID,
-			},
-			b3: B3{SingleHeader: true},
-			expected: &Supplier{
-				SingleHeader: "000000000000007b00000000000001c8-000000000000007b-0",
-			},
-		},
-		{
-			sc: SpanContext{
-				TraceID:    traceID,
-				SpanID:     spanID,
-				TraceFlags: FlagsUnused,
-			},
-			b3: B3{SingleHeader: true},
-			expected: &Supplier{
-				SingleHeader: "000000000000007b00000000000001c8-000000000000007b",
-			},
-		},
-		{
-			sc: SpanContext{
-				TraceID:    traceID,
-				SpanID:     spanID,
-				TraceFlags: FlagsSampled,
-			},
-			b3: B3{SingleHeader: true, SingleAndMultiHeader: true},
-			expected: &Supplier{
-				TraceIDHeader: traceIDStr,
-				SpanIDHeader:  spanIDStr,
-				SampledHeader: "1",
-				SingleHeader:  "000000000000007b00000000000001c8-000000000000007b-1",
-			},
-		},
-		{
-			sc: SpanContext{
-				TraceID:    traceID,
-				SpanID:     spanID,
-				TraceFlags: FlagsUnused,
-			},
-			b3: B3{},
-			expected: &Supplier{
-				TraceIDHeader: traceIDStr,
-				SpanIDHeader:  spanIDStr,
-			},
-		},
-		{
-			sc: SpanContext{
-				TraceID: traceID,
-				SpanID:  spanID,
-			},
-			b3: B3{},
-			expected: &Supplier{
-				TraceIDHeader: traceIDStr,
-				SpanIDHeader:  spanIDStr,
-				SampledHeader: "0",
-			},
-		},
-		{
-			sc: SpanContext{
-				TraceID:    traceID,
-				SpanID:     spanID,
-				TraceFlags: FlagsSampled,
-			},
-			b3: B3{},
-			expected: &Supplier{
-				TraceIDHeader: traceIDStr,
-				SpanIDHeader:  spanIDStr,
-				SampledHeader: "1",
-			},
-		},
-	}
-
-	for _, test := range tests {
-		ctx := ContextWithSpan(context.Background(), TestSpan{sc: test.sc})
-		actual := new(Supplier)
-		test.b3.Inject(ctx, actual)
-		assert.Equal(
-			t,
-			test.expected,
-			actual,
-			"B3: %#v, SpanContext: %#v", test.b3, test.sc,
-		)
-	}
-}
-
-func TestExtract(t *testing.T) {
-	tests := []struct {
-		traceID      string
-		spanID       string
-		parentSpanID string
-		sampled      string
-		flags        string
-		single       string
-		expected     SpanContext
-	}{
-		{
-			traceID:      "",
-			spanID:       "",
-			parentSpanID: "",
-			sampled:      "",
-			flags:        "",
-			single:       "",
-			expected:     empty,
-		},
-		{
-			traceID:      "",
-			spanID:       "",
-			parentSpanID: "",
-			sampled:      "",
-			flags:        "",
-			single:       "000000000000007b00000000000001c8-000000000000007b-1-00000000000001c8",
-			expected:     SpanContext{TraceID: traceID, SpanID: spanID, TraceFlags: FlagsSampled},
-		},
-		{
-			traceID:      traceIDStr,
-			spanID:       spanIDStr,
-			parentSpanID: "00000000000001c8",
-			sampled:      "1",
-			flags:        "",
-			single:       "",
-			expected:     SpanContext{TraceID: traceID, SpanID: spanID, TraceFlags: FlagsSampled},
-		},
-		{
-			traceID:      traceIDStr,
-			spanID:       spanIDStr,
-			parentSpanID: "00000000000001c8",
-			sampled:      "1",
-			flags:        "",
-			single:       "000000000000007b00000000000001c8-000000000000007b-1-00000000000001c8",
-			expected:     SpanContext{TraceID: traceID, SpanID: spanID, TraceFlags: FlagsSampled},
-		},
-	}
-
-	b3 := B3{}
-	for _, test := range tests {
-		ctx := context.Background()
-		supplier := &Supplier{
-			SingleHeader:       test.single,
-			DebugFlagHeader:    test.flags,
-			TraceIDHeader:      test.traceID,
-			SpanIDHeader:       test.spanID,
-			SampledHeader:      test.sampled,
-			ParentSpanIDHeader: test.parentSpanID,
-		}
-		info := []interface{}{
-			"trace ID: %q, span ID: %q, parent span ID: %q, sampled: %q, flags: %q, single: %q",
-			test.traceID,
-			test.spanID,
-			test.parentSpanID,
-			test.sampled,
-			test.flags,
-			test.single,
-		}
-		actual := RemoteSpanContextFromContext(b3.Extract(ctx, supplier))
-		assert.Equal(t, test.expected, actual, info...)
-	}
-}
 
 func TestExtractMultiple(t *testing.T) {
 	tests := []struct {
